refactor: move & rename main class
